added notifications, with timers in the fragment for workout sessions

+++ app/src/main/java/edu/rosehulman/bockkedummitrj/atomicaerobic/Interval.kt
@@ -4,5 +4,5 @@ data class Interval(
     var workout: String,
     var hour: Int,
     var minute : Int,
-    var duration: Int // in seconds always
+    var duration: Long // in seconds always
 )
